+++ classes/Form.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Grav\Plugin\Form;
 
 use ArrayAccess;