+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 1. [](#new)
     * Added support for custom form layouts
+    * Added Twig function `prepare_form_field()` to prepare form field array
+    * Added Twig function `include_form_field()` to get all include paths for the field type
 
 # v5.0.3
 ## 06/15/2021